Update README.md
Get Started with RAD Studio Guide

A Guide for Developers New to Delphi and C++Builder. This e-book is ideal for those new to RAD Studio, or those looking to use more of the current tools they have. This book covers the essentials from Installation and Setup; Understanding the IDE.

[Download Get Started with RAD Studio Guide](https://developer.team/delphi/35319-get-started-with-rad-studio-guide.html)